requirements.txt
# AI Clinical Notes and Patient Management Assistant
# Python Dependencies

# GUI Framework
ttkbootstrap==1.10.1

# Speech Recognition
openai-whisper==20231117
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.24.3

# LLM Integration
requests==2.31.0

# Document Export
python-docx==1.1.0
reportlab==4.0.7

# Encryption
cryptography==41.0.7

# Database (built-in with Python)
# sqlite3 is included in Python standard library

# Audio Processing
pydub==0.25.1

# Additional utilities
Pillow==10.1.0

# Development and Packaging (optional)
pyinstaller==6.3.0
